refactor: rs -> relativestrength로 풀어서 저장
