+++ apps/stock-api/src/test/java/com/momentum/application/insight/RsInsightServiceTest.java
@@ -3,10 +3,10 @@ package com.momentum.application.insight;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.momentum.domain.rs.Kospi;
-import com.momentum.domain.rs.KospiRelativeStrength;
-import com.momentum.domain.rs.KospiRelativeStrengthRepository;
-import com.momentum.domain.rs.KospiRepository;
+import com.momentum.domain.relativestrength.Kospi;
+import com.momentum.domain.relativestrength.KospiRelativeStrength;
+import com.momentum.domain.relativestrength.KospiRelativeStrengthRepository;
+import com.momentum.domain.relativestrength.KospiRepository;
 import com.momentum.domain.stock.Stock;
 import com.momentum.domain.stock.StockRegime;
 import com.momentum.domain.stock.StockRepository;
