+++ core/stock/src/main/java/com/momentum/domain/relativestrength/KospiRelativeStrengthRepository.java
@@ -1,4 +1,4 @@
-package com.momentum.domain.rs;
+package com.momentum.domain.relativestrength;
 
 import com.momentum.domain.stock.Stock;
 import java.util.List;
