+++ core/stock/src/main/java/com/momentum/infrastructure/rs/KospiJpaRepository.java
@@ -1,6 +1,6 @@
 package com.momentum.infrastructure.rs;
 
-import com.momentum.domain.rs.Kospi;
+import com.momentum.domain.relativestrength.Kospi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KospiJpaRepository extends JpaRepository<Kospi, Integer> {
